move server to a separate package
